# assets/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Go Template server.",
        "title": "Go Template API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/user/login": {
            "post": {
                "description": "登入一個已註冊的使用者",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "登入使用者",
                "parameters": [
                    {
                        "description": "使用者登入資訊",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers_user.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登入成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-template_internal_utils_response.SuccessResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "錯誤的請求",
                        "schema": {
                            "$ref": "#/definitions/go-template_internal_utils_response.ErrorResponseData"
                        }
                    },
                    "401": {
                        "description": "使用者不存在或密碼錯誤",
                        "schema": {
                            "$ref": "#/definitions/go-template_internal_utils_response.ErrorResponseData"
                        }
                    },
                    "500": {
                        "description": "系統錯誤",
                        "schema": {
                            "$ref": "#/definitions/go-template_internal_utils_response.ErrorResponseData"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "註冊一個新的使用者",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "註冊使用者",
                "parameters": [
                    {
                        "description": "使用者資料",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-template_internal_models.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "註冊成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-template_internal_utils_response.SuccessResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/go-template_internal_models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "錯誤的請求",
                        "schema": {
                            "$ref": "#/definitions/go-template_internal_utils_response.ErrorResponseData"
                        }
                    },
                    "500": {
                        "description": "系統錯誤",
                        "schema": {
                            "$ref": "#/definitions/go-template_internal_utils_response.ErrorResponseData"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "取得指定使用者的資訊",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "取得使用者資訊",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "取得成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-template_internal_utils_response.SuccessResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/go-template_internal_models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "使用者不存在",
                        "schema": {
                            "$ref": "#/definitions/go-template_internal_utils_response.ErrorResponseData"
                        }
                    },
                    "500": {
                        "description": "系統錯誤",
                        "schema": {
                            "$ref": "#/definitions/go-template_internal_utils_response.ErrorResponseData"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新指定使用者的資訊",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "更新使用者資訊",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "使用者 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "使用者資料",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-template_internal_models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-template_internal_utils_response.SuccessResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/go-template_internal_models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "錯誤的請求",
                        "schema": {
                            "$ref": "#/definitions/go-template_internal_utils_response.ErrorResponseData"
                        }
                    },
                    "404": {
                        "description": "使用者不存在",
                        "schema": {
                            "$ref": "#/definitions/go-template_internal_utils_response.ErrorResponseData"
                        }
                    },
                    "500": {
                        "description": "系統錯誤",
                        "schema": {
                            "$ref": "#/definitions/go-template_internal_utils_response.ErrorResponseData"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "刪除指定使用者",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "刪除使用者",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "使用者 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刪除成功",
                        "schema": {
                            "$ref": "#/definitions/go-template_internal_utils_response.SuccessResponseData"
                        }
                    },
                    "404": {
                        "description": "使用者不存在",
                        "schema": {
                            "$ref": "#/definitions/go-template_internal_utils_response.ErrorResponseData"
                        }
                    },
                    "500": {
                        "description": "系統錯誤",
                        "schema": {
                            "$ref": "#/definitions/go-template_internal_utils_response.ErrorResponseData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "go-template_internal_models.User": {
            "type": "object",
            "required": [
                "email",
                "username"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "description": "電子郵件",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_login": {
                    "description": "最後登入時間",
                    "type": "string"
                },
                "status": {
                    "description": "帳號狀態",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "description": "帳號名稱",
                    "type": "string"
                }
            }
        },
        "go-template_internal_utils_response.ErrorResponseData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "go-template_internal_utils_response.SuccessResponseData": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "internal_api_handlers_user.loginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}